selfcare-mcp-agent/pyproject.toml
[project]
name = "selfcare-mcp-agent"
version = "0.1.0"
description = "MCP server for the Self-Care Toolkit AI project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.15.0",
    "openai>=1.52.0",
    "openai-agents>=0.3.3",
    "python-dotenv>=1.1.1"
]
